Scrap builtin usage of Nextcloud OCC
The 'occ files:scan' command does not honour the --output=json option, it just prints the usual plaintext grid for interactive use. Since this is run via php, the process exit code is always zero regardless of occ errors, meaning we would have to capture stdout and parse the ascii grid to get the error count.

It is (in any case) far, far easier simply to run 'occ files:scan' via a cron job.

This also makes the upload endpoint independent of any particular storage product (such as Nextcloud) that might sit on top of the file system.